Resolve Eclipse/IntelliJ import order disagreement in style files.
The style files contained in `etc/` are not consistent between IDEs.
This commit modifies `etc/intellij-java-google-style.xml` to adhere to the import order specified in `etc/eclipseOrganizeImports.importorder`.
This seems a significant enough deviation from the original Google style guide to warrant a name change in the XML.
this closes #569