GEODE-2708: Declare ANTLR Dependency For antlr Configuration
If no version is set for the antlr plugin, it will default to 2.7.7. However,
this was causing an issue for Eclipse and other IDE because the plugin was
modifying the antlr configuration after the initial dependency resolution was
done.

- Added antlr dependency to the antlr configuration

Verified by importing project into Eclipse and also running build from
command-line