GEODE-3249: Validate internal client/server messages
This was merely a matter of changing the server to require the credentials
and changing the client to send credentials.  I removed the general overriding
of AbstractOp.processSecureBytes() because it made no sense.  If the server
sends a secure byte "part" in a message the client is obligated to process
it or the next message it sends will cause a security violation.

I've added a server-side property that folks can set to allow old clients
to continue to work.  This must be used to roll the servers forward to the
new version that contains this change.  Clients must then be rolled
forward & the servers can then be rolled once again without the property set.

The system property is
  geode.allow-internal-messages-without-credentials=true