whitelist wip
This addresses all of the problems in openBugs.txt, which now has no entries.

Several of the old Admin API classes couldn't be fixed, such as DistributionConfigSnapshot.
The old Admin API consequently will not work when a serialization filter is installed.

A handful of functions in extensions/geode-modules were causing problems in web server
testing and I modified them to be DataSerializable.

DistributionConfigImpl is still in excludedClasses.txt but it should not be there and
remain a Serializable.  We can move it back to sanctionedSerializables.txt when the
old Admin API is deleted.